% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPPC.R
\name{DPPC}
\alias{DPPC}
\title{Distributed projection principal component}
\usage{
DPPC(data, m, n1, K)
}
\arguments{
\item{data}{is a total data set}

\item{m}{is the number of principal component}

\item{n1}{is  the length of each data subset}

\item{K}{is the number of nodes}
}
\value{
Apro,pro,Sigmahathatpro
}
\description{
Distributed projection principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
DPPC(data_a,m=3,n1=128,K=2)
}
