% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FanPC.R
\name{FanPC}
\alias{FanPC}
\title{Apply the FanPC method to the Laplace factor model}
\usage{
FanPC(data, m)
}
\arguments{
\item{data}{A matrix of input data.}

\item{m}{is the number of principal component}
}
\value{
AF,DF,SigmahatF
}
\description{
This function performs Factor Analysis via Principal Component (FanPC) on a given data set. It calculates the estimated factor loading matrix (AF), specific variance matrix (DF), and the mean squared errors.
}
\examples{
library(LaplacesDemon)
library(MASS)
n=1000
p=10
m=5
mu=t(matrix(rep(runif(p,0,1000),n),p,n))
mu0=as.matrix(runif(m,0))
sigma0=diag(runif(m,1))
F=matrix(mvrnorm(n,mu0,sigma0),nrow=n)
A=matrix(runif(p*m,-1,1),nrow=p)
lanor <- rlaplace(n*p,0,1)
epsilon=matrix(lanor,nrow=n)
D=diag(t(epsilon)\%*\%epsilon)
data=mu+F\%*\%t(A)+epsilon
results <- FanPC(data, m)
print(results)
}
