% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlaplace.R
\name{rlaplace}
\alias{rlaplace}
\title{Generate Random Variables from the Laplace Distribution}
\usage{
rlaplace(n, b = 1)
}
\arguments{
\item{n}{An integer specifying the number of observations to generate.}

\item{b}{A positive numeric value specifying the scale parameter of the
Laplace distribution. The default value is \code{1}.}
}
\value{
A numeric vector of length \code{n} containing random observations from
the Laplace distribution.
}
\description{
Generates random samples from the Laplace (double exponential)
distribution with location parameter 0 and scale parameter \code{b}.

The probability density function is

\deqn{
f(x)=\frac{1}{2b}\exp\left(-\frac{|x|}{b}\right),
}

where \eqn{b>0} is the scale parameter.

Random numbers are generated using the inverse transform method.
}
\details{
If \eqn{X \sim Laplace(0,b)}, then

\deqn{
E(X)=0,
}

and

\deqn{
Var(X)=2b^2.
}

This function is used internally by the Laplace Factor Model (LFM)
package for generating Laplace-distributed error terms.
}
\examples{
set.seed(123)

x <- rlaplace(1000)

mean(x)
var(x)

hist(x,
     breaks = 30,
     main = "Laplace Random Sample",
     xlab = "x")

}
